Fix middle-snake algorithm for negative delta
The middle snake algorithm is supposed to check for overlapping snakes
either in forward mode (when delta is odd) or backward mode (when delta
is even). However the check in forward mode only checked for "delta % 2
=== 1" which does evaluate to false when delta is negative leading to
wrong results.

Because delta does not change during the method call, a new boolean
checkBwSnake is set to delta % 2 === 0 at the beginning of the method
which. Beside fixing the mentioned bug, this has the side effect, that
the condition is only evaluated once and also enhances code clarity.